+++ js/main.js
@@ -1,4 +1,3 @@
-import { createOffers } from './utils/create-offer.js';
-import { OFFER_COUNT } from './data.js';
+import { createOffers } from './data.js';
 
-createOffers(OFFER_COUNT);
+createOffers();
